stackon/pyproject.toml
[tool.poetry]
name = "stackon"
version = "0.1.0"
description = ""
authors = ["Dedalum <ha.lagrange9000@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
paho-mqtt = "^1.5.1"
gevent = "^21.1.2"

[tool.poetry.dev-dependencies]
matplotlib = "^3.4.2"
memory-profiler = "^0.58.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
